requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
# Gradio 4.44 json_schema breaks on pydantic 2.12+ (State / Chatbot api_info)
pydantic>=2.0.0,<2.11.0
gradio>=4.44,<5
huggingface_hub>=0.23.0,<1.0
